backend/requirements.txt
django
psycopg2-binary
djangorestframework
django-cors-headers
python-dotenv
django-allauth
requests